Create chap_8_LFS_11.2r-systemd_part2.sh
This is the second part.  All the previous requirements are still necessary, and obviously should be run after part 1 has completed...
IF it has completed successfully... 

Please read the readme and the introduction to part1.  Both of these scripts assume the reader has successfully built through Chapter 7 and knows how to bind the host kernel process and chroot.   Good luck.