Update chap_8_LFS_11.2r-systemd_part1.sh
Commented out :
make localedata/install-locales
in glibc.

The  book in both the release and current branch versions seem to indicate it is optional: "Alternatively, install all locales listed in the glibc-2.36/localedata/SUPPORTED file (it includes every locale listed above and many more) at once with the following time-consuming command:
make localedata/install-locales 
Not sure why its not working for me.... is it a typo?  For Languages not listed in the long list this could be an issue.